make-bcache: initialize sb block size with physical_block_size
Usually logical_block_size is 512B, when a cache dev has a 512B
sb block size, sometimes performance will be poor if an I/O is aligned
to 512B but not aligned to 4K(the SSD phsycial block size). So we want
to use physical_block_size here to avoid performance degradation.

Signed-off-by: Caspar Zhang <jinli.zjl@alibaba-inc.com>